+++ src/lib/client.ts
@@ -26,6 +26,7 @@ import { resolveHost, getVersion } from './utils';
 import { Upload, InputFile, UploadOptions, StoreUploadOptions, UploadTags } from './api/upload';
 import { preview, PreviewOptions } from './api/preview';
 import { CloudClient } from './api/cloud';
+import { Prefetch, PrefetchResponse, PrefetchOptions } from './api/prefetch';
 import { StoreParams } from './filelink';
 
 import { picker, PickerInstance, PickerOptions } from './picker';
@@ -39,6 +40,7 @@ export interface Session {
   cname?: string;
   policy?: string;
   signature?: string;
+  prefetch?: PrefetchResponse;
 }
 
 export interface Security {
@@ -92,8 +94,9 @@ export interface ClientOptions {
  * ```
  */
 export class Client extends EventEmitter {
-  session: Session;
+  public session: Session;
   private cloud: CloudClient;
+  private prefetchInstance: Prefetch;
 
   private forwardErrors: boolean = false;
 
@@ -125,7 +128,17 @@ export class Client extends EventEmitter {
       this.setCname(cname);
     }
 
+    this.prefetchInstance = new Prefetch(this.session);
     this.cloud = new CloudClient(this.session, options);
+  }
+
+  /**
+   * Make basic prefetch request to check permissions
+   *
+   * @param params
+   */
+  prefetch(params: PrefetchOptions) {
+    return this.prefetchInstance.getConfig(params);
   }
 
   /**