+++ src/lib/picker.ts
@@ -15,10 +15,11 @@
  * limitations under the License.
  */
 
-import { loadModule, knownModuleIds } from 'filestack-loader';
+import { loadModule, FILESTACK_MODULES } from '@filestack/loader';
 import { FilestackError, FilestackErrorType } from './../filestack_error';
 import { Client } from './client';
 import { FSProgressEvent, UploadOptions, WorkflowConfig } from './api/upload/types';
+import { UploadTags } from './api/upload/file';
 import { getValidator, PickerParamsSchema } from './../schema';
 
 export interface PickerInstance {
@@ -662,6 +663,14 @@ export interface PickerOptions {
    * It can override global objects like console, error etc. Defaults to `true`.
    */
   useSentryBreadcrumbs?: boolean;
+
+  /**
+   * User upload tags
+   *
+   * @type {UploadTags}
+   * @memberof PickerOptions
+   */
+  uploadTags?: UploadTags;
 }
 
 export interface PickerCropOptions {
@@ -704,9 +713,9 @@ export interface PickerTransformationOptions {
 class PickerLoader {
 
   private _initialized: Promise<PickerInstance>;
-
   constructor(client: Client, options?: PickerOptions) {
     const validateRes = getValidator(PickerParamsSchema)(options);
+
     if (validateRes.errors.length) {
       throw new FilestackError(`Invalid picker params`, validateRes.errors, FilestackErrorType.VALIDATION);
     }
@@ -736,7 +745,7 @@ class PickerLoader {
 
   private async loadModule(client: Client, options?: PickerOptions): Promise<PickerInstance> {
     const { session: { urls: { pickerUrl: url } } } = client;
-    const Picker = await loadModule(url, knownModuleIds.picker);
+    const Picker = await loadModule(FILESTACK_MODULES.PICKER, url);
     return new Picker(client, options);
   }
 }